@@ -10,7 +10,6 @@ classifiers = [
     "Intended Audience :: Developers",
     "Operating System :: OS Independent",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -20,7 +19,7 @@ name = "nomad-measurements"
 dynamic = ["version"]
 description = "A plugin for NOMAD containing base sections for measurements."
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 authors = [
     { name = "Andrea Albino" },
     { name = "Sebastian Brückner" },
@@ -35,7 +34,7 @@ maintainers = [
 ]
 license = { file = "LICENSE" }
 dependencies = [
-    "nomad-lab>=1.3.6", 
+    "nomad-lab>=1.3.14",
     "xmltodict==0.13.0",
     "fairmat-readers-xrd>=0.0.3",
     "nomad-material-processing",
